[Tests] Refactor ExecutorTest utilities into RuntimeUtils
Description:
This commit moves the TestDeviceManager, ExecutorTestBuilder and
ExecutorTest into RuntimeTestUtils so that they can be reused by
ExecutorBenchmark. The benchmarks will also need to specify some DAG
topology and create all of the necessary scaffolding (e.g. RuntimeBundles,
intermediate ExecutionContexts, etc.) in order to properly measure
overhead from tensor cloning, late notification of failures, etc.

Testing:
All unit tests still pass.
